Propagate errors through SignFile, GenerateBundle.
This allows the application to try and sign as many files as it can, as
well as generate as many bundles as possible, but still exit with an
exit code different from 0. SignFile would exit the application before
signing other files, while GenerateBundle would fail without reporting
it in the exit code.