Ignore EOF errors in CheckMSDos
CheckMSDos attempts to read the first two bytes of a file to determine
if it is an MS-DOS executable. If a file (for whatever reason) is shorter,
attempting to read this will result in an `io.EOF` error.

The error can be safely ignore because such an error indicates that the file
could be read successfully, it just doesn't contain enough bytes.