Fix CI post-CVE-2022-24765
Due to CVE-2022-24765, git(1) won't operate on a repository owned by a
separate user. sbctl's tests run inside a container, and the user does
not match the owner of the repository, so this broke tests.

Also tried using actions/checkout@v3 and actions/checkout@v3.0.1, but
neither of those work.