Read output file from db if output not specified
The mkinitcpio hook runs `sbctl sign "$IMAGE_FILE"` and has no way to
know the output filepath for the corresponding image file. The results
in signing in place which may break booting for those who with different
locations for signed and unsigned UKIs.

The sign command should check if the given file is in the database and
use the saved output filepath.